Added Scaling and made Matrix a class
These are two milestones and probably should have been done in
different commits at the very least.

Scaling hasn't been tested for its ability to scale yet, only
that it doesn't interfere with computations. Scaling in the matrix
class will scale properly, between 1e-10 and -1e-10 and greater and
less than 1e10 and -1e10 respectively. This is not yet the case
in the kernel.

Matrix is now a class with all sorts of classy stuff. Naming and
behavior still needs to be refactored or implemented so everything
behaves in a more intuitive manner.