Found two problems:
1. Padding isn't updating the bounding offset
2. For whatever reason there is data duplication in the device buffers

They may be related, so fix the first one first. The offset issue I can fix
in one of two ways:

1. heuristic changes to the offset during the padding process. This is
extremely bittle and requires the user to know the dimensions of the array.
2. Implement bracket and slice operators, or some other method of bounding
that replaces the singular offset with a start and end index in each
dimension.
3. Replace the single offset with row and col offsets.

The second one would be a lot of work but could have some serious advantages.
The third sounds about as good and would be less confusing, perhaps.