Multiply working again
The problem was that the overwrite switch was defaulting to off,
which meant that everything going into C was multiplied by 0
before writing. This, unsurprisingly, produced a lot of zeros.

To resolve the problem in HyPhy, the parent node values need to be
initialized to 1 OR the overwrite node needs to be switched
depending on the value saved in "tagged internals" (or w/e).