Working double tests
Everything is working fine, but if using a double matrix as a
source matrix you're not going to get array reuse down the line.

The best way to resolve this would probably be to have a set_m()
method that takes another Matix object, but that would have the
Matrix class exposed to the user. Instead there will probably
have to be a bit more work done in the set_m methods to detect
when two matrices are set to the same data. It then might be a
good idea to make a MatrixData class to store all the different
versions of whatever has been passed in.

That being said, it would be a good idea to keep around a double
source matrix as that is what is the expected behavior if a float
source matrix is passed in.